(Update) model and test data
Update model to use correct number of decimals for price. Update test data to be correct

+++ server/db/models/Product.js
@@ -22,7 +22,7 @@ module.exports = (sequelize, DataTypes) => {
       type: DataTypes.INTEGER
     },
     price: {
-      type: DataTypes.INTEGER
+      type: DataTypes.DECIMAL(10,2)
     },
     image: {
       type: DataTypes.TEXT
